Removed "fromItem" from ShowSpecial()    This affects two groups of items:      battle items where it sounds weird:       staff of fire       staff of lightning       runesword
"Item can cast a fireball in battle at a skill level of 3. This ability
 only affects the possessor of the item."
(are you fireballing yourself?)

   and shield items which sound contradictory:
      amulet of invulnerability
      amulet of protection
      shieldstone

"This ability only affects the possessor of the item. This ability provides a
shield against all all attacks against the entire army at a level equal to the
skill level of the ability."
(affects only the possessor, but shield the entire army?)